requirements.txt
numpy
librosa
tqdm
hydra-core
pytest
scikit-learn
omegaconf
soundfile
pandas
wandb
joblib
lxml
future-fstrings
